% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carte.R
\name{carte}
\alias{carte}
\title{carte}
\usage{
carte()
}
\value{
carte djelfa
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\examples{
carte
}
