% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carte_w_c_p.R
\name{carte_w_c_p}
\alias{carte_w_c_p}
\title{Title  carte_w_c_p}
\usage{
carte_w_c_p(df, wilaya)
}
\arguments{
\item{wilaya}{wilaya}

\item{variables}{df}
}
\value{
carte
}
\description{
Title  carte_w_c_p
}
\examples{
yahiatiba::carte_w_c_p (variables,wilaya)
}
