% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carte_w_c_pv.R
\name{carte_w_c_pv}
\alias{carte_w_c_pv}
\title{Title carte_w_c_pv}
\usage{
carte_w_c_pv(df, wilaya_x)
}
\arguments{
\item{df}{df}

\item{wilaya_x}{wilaya_x}
}
\value{
value
}
\description{
Title carte_w_c_pv
}
\examples{
yahiatiba::carte_w_c_pv
}
