% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declaration_ta.R
\name{declaration_ta}
\alias{declaration_ta}
\title{Title declaration_ta}
\usage{
declaration_ta(DF)
}
\arguments{
\item{DF}{value}
}
\value{
value
}
\description{
Title declaration_ta
}
\examples{
value
}
