% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declaration_ta1.R
\name{declaration_ta1}
\alias{declaration_ta1}
\title{Title declaration_ta1}
\usage{
declaration_ta1(DF, TO, BY)
}
\arguments{
\item{DF}{value}

\item{TO}{value}

\item{BY}{value}
}
\value{
value
}
\description{
Title declaration_ta1
}
\examples{
value
}
