% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decpta_commune.R
\name{decpta_commune}
\alias{decpta_commune}
\title{Title decpta_commune}
\usage{
decpta_commune(DF)
}
\arguments{
\item{DF}{value}
}
\value{
value
}
\description{
Title decpta_commune
}
\examples{
value
}
