% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decpta_pathologie.R
\name{decpta_pathologie}
\alias{decpta_pathologie}
\title{Title decpta_pathologie}
\usage{
decpta_pathologie(DF)
}
\arguments{
\item{DF}{value}
}
\value{
value
}
\description{
Title decpta_pathologie
}
\examples{
value
}
