% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drh_grade1.R
\name{drh_grade1}
\alias{drh_grade1}
\title{Title drh_grade1}
\usage{
drh_grade1(drh, wilaya)
}
\arguments{
\item{drh}{value}

\item{wilaya}{value}
}
\value{
value
}
\description{
Title drh_grade1
}
\examples{
value
}
