% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drh_specialiste_str.R
\name{drh_specialiste_str}
\alias{drh_specialiste_str}
\title{Title drh_specialiste_str}
\usage{
drh_specialiste_str(drh)
}
\arguments{
\item{drh}{value}
}
\value{
value
}
\description{
Title drh_specialiste_str
}
\examples{
value
}
