% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drh_drh_specialiste_str1.R
\name{drh_specialiste_str1}
\alias{drh_specialiste_str1}
\title{Title drh_specialiste_str1}
\usage{
drh_specialiste_str1(drh)
}
\arguments{
\item{drh}{value}
}
\value{
value
}
\description{
Title drh_specialiste_str1
}
\examples{
value
}
