% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drh_specialite.R
\name{drh_specialite}
\alias{drh_specialite}
\title{Title drh_specialite}
\usage{
drh_specialite(drh)
}
\arguments{
\item{drh}{value}
}
\value{
value
}
\description{
Title drh_specialite
}
\examples{
value
}
