% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat_aindefla.R
\name{mat_aindefla}
\alias{mat_aindefla}
\title{Title mat_aindefla}
\usage{
mat_aindefla(mat, wil)
}
\arguments{
\item{mat}{value}

\item{wil}{value}
}
\value{
value
}
\description{
Title mat_aindefla
}
\examples{
mat_aindefla
}
