% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat_chlef.R
\name{mat_chlef}
\alias{mat_chlef}
\title{Title mat_chlef}
\usage{
mat_chlef(mat, wil)
}
\arguments{
\item{mat}{value}

\item{wil}{value}
}
\value{
value
}
\description{
Title mat_chlef
}
\examples{
mat_chlef
}
