% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat_creance.R
\name{mat_creance}
\alias{mat_creance}
\title{mat creance}
\usage{
mat_creance(wil, creance)
}
\arguments{
\item{wil}{value}

\item{creance}{value}
}
\value{
value
}
\description{
mat creance
}
\examples{
mat_creance (wil,creance)
}
