% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat_msila.R
\name{mat_msila}
\alias{mat_msila}
\title{Title mat_msila}
\usage{
mat_msila(mat, wil)
}
\arguments{
\item{mat}{value}

\item{wil}{value}
}
\value{
value
}
\description{
Title mat_msila
}
\examples{
mat_msila
}
