% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat_region.R
\name{mat_region}
\alias{mat_region}
\title{Title mat_region}
\usage{
mat_region(mat, wil)
}
\arguments{
\item{mat}{value}

\item{wil}{value}
}
\value{
value
}
\description{
Title mat_region
}
\examples{
mat_region
}
