% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat_repartition.R
\name{mat_repartition}
\alias{mat_repartition}
\title{mat_repartition}
\usage{
mat_repartition(mat, wil, str)
}
\arguments{
\item{mat}{value}

\item{wil}{value}

\item{str}{value}
}
\value{
value
}
\description{
mat_repartition
}
\examples{
mat_repartition (mat,wil,str)
}
