% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yahiatiba-package.R
\docType{data}
\name{mdo}
\alias{mdo}
\title{Points aleatoires}
\format{
Une matrice contenant 10 points designes par les
coordonnees suivantes.
\describe{
\item{\code{X}}{ coordonnee en X }
\item{\code{Y}}{ coordonnee en Y }
}
}
\usage{
mdo
}
\description{
Coordonnes en deux dimensions de 10 points aleatoires.
}
\examples{
dist_manhattan(point1 = points[4, ], point2 = points[8, ])

# Note : Ces donnees ont ete crees par les instructions suivantes
points <- matrix(sample(1:10, size = 20, replace = TRUE),
nrow = 10, ncol = 2)
colnames(points) <- c("X", "Y")
}
\keyword{datasets}
