% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ext_data.R
\name{read_ext_data}
\alias{read_ext_data}
\title{read ext data create path to}
\usage{
read_ext_data(pkgname, path = NULL)
}
\arguments{
\item{pkgname}{value}

\item{path}{value}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{    value
}\if{html}{\out{</div>}}
}
\description{
read ext data create path to
}
\examples{
  value
}
