% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shp_dz}
\alias{shp_dz}
\title{shp_dz shp data}
\format{
\subsection{\code{shp_dz}}{

A data frame with 1 rows and 3 columns:
\describe{
\item{pay}{Algerie }
\item{year}{2024}
}
}
}
\source{
\url{https://tibaredha/}
}
\usage{
shp_dz
}
\description{
A subset of data from the shp_dz
}
\details{
\code{tibaredha \link{tibaredha}}
}
\keyword{datasets}
