% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shp_dzc}
\alias{shp_dzc}
\title{shp_dzc shp data}
\format{
\subsection{\code{shp_dzc}}{

A data frame with 1504 rows and 4 columns:
\describe{
\item{pay}{Algerie }
\item{wilaya}{48 wilayas}
\item{commune}{1504 communes}
\item{year}{2024}
}
}
}
\source{
\url{https://tibaredha/}
}
\usage{
shp_dzc
}
\description{
A subset of data from the shp_dzc
}
\details{
\code{tibaredha \link{tibaredha}}
}
\keyword{datasets}
