% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shp_dzrc}
\alias{shp_dzrc}
\title{shp_dzrc shp data}
\format{
\subsection{\code{shp_dzrc}}{

A data frame with 242 rows and 4 columns:
\describe{
\item{pay}{Algerie }
\item{wilaya}{06 wilayas}
\item{commune}{242 communes}
\item{year}{2024}
}
}
}
\source{
\url{https://tibaredha/}
}
\usage{
shp_dzrc
}
\description{
A subset of data from the shp_dzrc
}
\details{
\code{tibaredha \link{tibaredha}}
}
\keyword{datasets}
