% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shp_dzrw}
\alias{shp_dzrw}
\title{shp_dzrw shp data}
\format{
\subsection{\code{shp_dzrw}}{

A data frame with 6 rows and 3 columns:
\describe{
\item{pay}{Algerie }
\item{wilaya}{06 wilayas}
\item{year}{2024}
}
}
}
\source{
\url{https://tibaredha/}
}
\usage{
shp_dzrw
}
\description{
A subset of data from the shp_dzrw
}
\details{
\code{tibaredha \link{tibaredha}}
}
\keyword{datasets}
