% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shp_dzw}
\alias{shp_dzw}
\title{shp_dzw shp data}
\format{
\subsection{\code{shp_dzw}}{

A data frame with 48 rows and 3 columns:
\describe{
\item{pay}{Algerie }
\item{wilaya}{48 wilayas}
\item{year}{2024}
}
}
}
\source{
\url{https://tibaredha/}
}
\usage{
shp_dzw
}
\description{
A subset of data from the shp_dzw
}
\details{
\code{tibaredha \link{tibaredha}}
}
\keyword{datasets}
