% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_lit1_e_x.R
\name{str_lit1_e_x}
\alias{str_lit1_e_x}
\title{Title str.lit1_e_x}
\usage{
str_lit1_e_x(DF, xeta)
}
\arguments{
\item{DF}{value}

\item{xeta}{vualue}
}
\value{
value
}
\description{
Title str.lit1_e_x
}
\examples{
value
}
