% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_lit_wil.R
\name{str_lit_wil}
\alias{str_lit_wil}
\title{Title str.lit_wil}
\usage{
str_lit_wil(DF)
}
\arguments{
\item{DF}{value}
}
\value{
value
}
\description{
Title str.lit_wil
}
\examples{
value
}
\references{
\url{https://en.wikipedia.org/wiki/List_of_Crayola_crayon_colors}
}
\seealso{
\code{\link{brocolors}}
}
\author{
TIBA   Rdha, \email{tibaredha@yahoo.fr}
}
\keyword{...}
\keyword{hplot}
