% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_ser_lit.R
\name{str_ser_lit}
\alias{str_ser_lit}
\title{Title str.ser_lit}
\usage{
str_ser_lit(DF, XSERVICE)
}
\arguments{
\item{DF}{value}

\item{XSERVICE}{value}
}
\value{
value
}
\description{
Title str.ser_lit
}
\examples{
value
}
