% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wc_algeria}
\alias{wc_algeria}
\title{wc_algeria sh data}
\format{
\subsection{\code{wc_algeria}}{

A data frame with 224 rows and 4 columns:
\describe{
\item{pay}{Algerie }
\item{wilaya}{48 wilayas}
\item{year}{2024}
}
}
}
\source{
\url{https://tibaredha/}
}
\usage{
wc_algeria
}
\description{
A subset of data from the wca
}
\keyword{datasets}
