% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wc_region}
\alias{wc_region}
\title{wc_region sh data}
\format{
\subsection{\code{wc_region}}{

A data frame with 224 rows and 4 columns:
\describe{
\item{pay}{Algerie }
\item{region}{Médéa}
\item{year}{2024}
}
}
}
\source{
\url{https://tibaredha/}
}
\usage{
wc_region
}
\description{
A subset of data from the wca
}
\keyword{datasets}
