% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdc_communes.R
\name{wdc_communes}
\alias{wdc_communes}
\title{Title wdc_communes}
\usage{
wdc_communes(df, variables)
}
\arguments{
\item{df}{df}

\item{variables}{wilaya}
}
\value{
table
}
\description{
Title wdc_communes
}
\examples{
yahiatiba::wdc_communes(wdc,"Djelfa")
}
