% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdc_dairas.R
\name{wdc_dairas}
\alias{wdc_dairas}
\title{Title wdc_dairas}
\usage{
wdc_dairas(df, variables)
}
\arguments{
\item{df}{df}

\item{variables}{Wilaya}
}
\value{
table
}
\description{
Title wdc_dairas
}
\examples{
yahiatiba::wdc_dairas(wdc,"Djelfa")
}
