% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yahiatiba-package.R
\docType{package}
\name{yahiatiba-package}
\alias{yahiatiba}
\alias{yahiatiba-package}
\title{yahiatiba: maladie a declaration obligatoire}
\description{
maladie a declaration obligatoire.
}
\author{
\strong{Maintainer}: Redha TIBA \email{prenom.nom@insee.fr}

Authors:
\itemize{
  \item Redha TIBA
  \item azzedine BOULAOUCHE
}

Other contributors:
\itemize{
  \item xx yyy [contributor]
  \item  Institut national de la statistique et des études économiques [copyright holder]
}

}
\keyword{internal}
