% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yahiatiba.R
\docType{package}
\name{yahiatiba}
\alias{yahiatiba}
\title{yahiatiba: A package for computating the notorious bar statistic}
\description{
The yahiatiba package provides three categories of important functions:
foo, bar and baz.
}
\section{Foo functions}{

The foo functions ...
}

\author{
\strong{Maintainer}: Redha TIBA \email{prenom.nom@insee.fr}

Authors:
\itemize{
  \item Redha TIBA
  \item azzedine BOULAOUCHE
}

Other contributors:
\itemize{
  \item xx yyy [contributor]
  \item  Institut national de la statistique et des études économiques [copyright holder]
}

}
