% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat_laghouat.R
\name{mat_laghouat}
\alias{mat_laghouat}
\title{mat_laghouat}
\usage{
mat_laghouat(mat, wil)
}
\arguments{
\item{mat}{value}

\item{wil}{value}
}
\value{
value
}
\description{
mat_laghouat
}
\examples{
mat_laghouat
}
