% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yahiatiba.R
\docType{package}
\name{yahiatiba}
\alias{yahiatiba}
\title{yahiatiba: A package for computating the notorious bar statistic}
\description{
The yahiatiba package provides three categories of important functions:
foo, bar and baz.
}
\section{yahiatiba functions}{

The yahiatiba functions ...
}

